# documention/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/national/programme/create": {
      "post": {
        "operationId": "ProgrammeController_addProgramme",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/addNDCAction": {
      "post": {
        "operationId": "ProgrammeController_addNDCAction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NDCActionDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/addDocument": {
      "post": {
        "operationId": "ProgrammeController_addDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeDocumentDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/docAction": {
      "post": {
        "operationId": "ProgrammeController_docAction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DocumentAction" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/queryDocs": {
      "post": {
        "operationId": "ProgrammeController_queryDocuments",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/queryNdcActions": {
      "post": {
        "operationId": "ProgrammeController_queryNdcActions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/queryNdcActions/download": {
      "post": {
        "operationId": "ProgrammeController_getNdcDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/query": {
      "post": {
        "operationId": "ProgrammeController_getAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/download": {
      "post": {
        "operationId": "ProgrammeController_getDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/getHistory": {
      "get": {
        "operationId": "ProgrammeController_getHistory",
        "parameters": [
          {
            "name": "programmeId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }, { "api_key": [] }]
      }
    },
    "/national/programme/updateConfigs": {
      "post": {
        "operationId": "ProgrammeController_updateConfigs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ConstantUpdateDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/authorize": {
      "put": {
        "operationId": "ProgrammeController_programmeApprove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeApprove" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/issue": {
      "put": {
        "operationId": "ProgrammeController_programmeIssue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeMitigationIssue"
              }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/reject": {
      "put": {
        "operationId": "ProgrammeController_programmeReject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeReject" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/retire": {
      "put": {
        "operationId": "ProgrammeController_programmeRetire",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeRetire" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/certify": {
      "put": {
        "operationId": "ProgrammeController_programmeCertify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeCertify" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/revoke": {
      "put": {
        "operationId": "ProgrammeController_programmeRevoke",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProgrammeRevoke" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transferRequest": {
      "post": {
        "operationId": "ProgrammeController_transferRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeTransferRequest"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transferApprove": {
      "post": {
        "operationId": "ProgrammeController_transferApprove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeTransferApprove"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transferReject": {
      "post": {
        "operationId": "ProgrammeController_transferReject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeTransferReject"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transferCancel": {
      "post": {
        "operationId": "ProgrammeController_transferCancel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgrammeTransferCancel"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transferQuery": {
      "post": {
        "operationId": "ProgrammeController_queryUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transfers/download": {
      "post": {
        "operationId": "ProgrammeController_getTransfersDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/transfersByProgrammeId": {
      "get": {
        "operationId": "ProgrammeController_transfersByProgrammeId",
        "parameters": [
          {
            "name": "programmeId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/addInvestment": {
      "post": {
        "operationId": "ProgrammeController_addInvestment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestmentRequestDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/investmentApprove": {
      "post": {
        "operationId": "ProgrammeController_investmentApprove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestmentApprove" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/investmentReject": {
      "post": {
        "operationId": "ProgrammeController_investmentReject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestmentReject" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/investmentCancel": {
      "post": {
        "operationId": "ProgrammeController_investmentCancel",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InvestmentCancel" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/investmentQuery": {
      "post": {
        "operationId": "ProgrammeController_queryInvestmentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/investments/download": {
      "post": {
        "operationId": "ProgrammeController_getInvestmentsDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/programme/queryNdcDetailsPeriod": {
      "get": {
        "operationId": "ProgrammeController_getNdcDetailsPeriods",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/programme/addNdcDetailsPeriod": {
      "post": {
        "operationId": "ProgrammeController_addNdcDetailsPeriod",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NdcDetailsPeriodDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/programme/deleteNdcDetailsPeriod": {
      "post": {
        "operationId": "ProgrammeController_deleteNdcDetailsPeriod",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "number" } } }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/programme/finalizeNdcDetailsPeriod": {
      "post": {
        "operationId": "ProgrammeController_finalizeNdcDetailsPeriod",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "number" } } }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/programme/queryNdcDetailsAction": {
      "get": {
        "operationId": "ProgrammeController_getNdcDetailActions",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 1 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "National API",
    "description": "RESTful Web API Documentation.",
    "version": "0.5",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "api_key": { "type": "apiKey", "in": "header", "name": "api_key" }
    },
    "schemas": {
...
      "ProgrammeDocumentDto": {
        "type": "object",
        "properties": {
          "data": { "type": "string" },
          "programmeId": { "type": "string" },
          "txTime": { "type": "number" },
          "status": { "type": "string" },
          "externalId": { "type": "string" },
          "type": {
            "type": "string",
            "enum": [
              "0",
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "10",
              "11",
              "12",
              "13",
              "14",
              "15",
              "16",
              "17",
              "18",
              "19",
              "20",
              "21",
              "22",
              "23",
              "24",
              "25",
              "26",
              "27",
              "28",
              "29",
              "30",
              "31",
              "32",
              "33",
              "34",
              "35",
              "36",
              "37",
              "38"
            ]
          },
          "actionId": { "type": "string" },
          "certifierTaxId": { "type": "string" }
        },
        "required": ["data", "txTime", "status", "type"]
      },
...
    }
  }
}